+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a flatpak job in the CI
 
 ### Changed
+- main.py slimmed down by about 1000 lines
 
 ### Fixed